@@ -26,33 +25,32 @@
     "faker": "^6.6.6",
     "formik": "^2.2.9",
     "hamburger-react": "^2.5.0",
+    "history": "^5.3.0",
     "html-webpack-plugin": "^5.5.0",
     "identity-obj-proxy": "3.0.0",
     "immer": "^9.0.15",
     "jest": "^28.1.3",
     "localforage": "^1.10.0",
-    "mui-datatables": "^4.2.2",
+    "mui-datatables": "4.3.0",
     "node-polyfill-webpack-plugin": "^2.0.1",
-    "node-sass": "^7.0.1",
     "os-browserify": "^0.3.0",
+    "potrace": "2.1.8",
     "prettier": "^2.5.1",
     "query-string": "^7.1.1",
-    "react": "17.0.2",
+    "react": "^18.2.0",
     "react-app-rewired": "^2.2.1",
-    "react-card-flip": "^1.1.5",
+    "react-card-flip": "^1.2.3",
     "react-device-detect": "^2.2.2",
-    "react-dom": "17.0.2",
+    "react-dom": "^18.2.0",
     "react-draggable": "^4.4.5",
     "react-helmet": "^6.1.0",
-    "react-hot-loader": "4.13.0",
     "react-modal": "3.15.1",
     "react-number-format": "4.9.3",
     "react-redux": "^8.0.2",
-    "react-resizable": "1.11.0",
-    "react-router": "5.1.2",
-    "react-router-dom": "5.1.2",
+    "react-resizable": "^3.0.5",
+    "react-router-dom": "^6.30.1",
     "react-scripts": "5.0.0",
-    "react-select": "4.3.1",
+    "react-select": "^5.9.0",
     "react-spinners": "^0.13.4",
     "react-spring": "^9.5.2",
     "react-switch": "^7.0.0",
