@@ -1,10 +1,12 @@
 import React from 'react';
-import ReactDOM from 'react-dom';
+import ReactDOM from 'react-dom/client';
 import { BrowserRouter } from 'react-router-dom';
 import App from './components/app';
 import ProviderWithRouter from './components/providerWithRouter';
 
-ReactDOM.render(
+const root = ReactDOM.createRoot(document.getElementById('root')!);
+
+root.render(
   <React.StrictMode>
     <BrowserRouter basename="/project-r">
       <ProviderWithRouter>
@@ -12,5 +14,4 @@ ReactDOM.render(
       </ProviderWithRouter>
     </BrowserRouter>
   </React.StrictMode>,
-  document.getElementById('root'),
 );
