@@ -1,7 +1,7 @@
 /* eslint-disable react-hooks/exhaustive-deps */
 import { useEffect, useState, useRef } from 'react';
 import { useDispatch } from 'react-redux';
-import Editor, { EditorDidMount } from '@monaco-editor/react';
+import Editor, { OnMount } from '@monaco-editor/react';
 import prettier from 'prettier';
 import parser from 'prettier/parser-babel';
 import { useCumulativeCode } from '../../hooks/useCumulativeCode';
@@ -12,16 +12,12 @@ import LoadingSpinner from '../../elements/loadingSpinner';
 import Resizable from '../../elements/resizable';
 import Icon from '../../elements/icon';
 import scssObj from './_CodeCell.scss';
-import { HTML } from '../../utils/consts';
+import { getHTML } from '../../utils/consts';
 import { copyText } from '../../redux/me';
 
 interface Props {
   cell: Cell;
-  bundle?: {
-    processing: boolean;
-    code: string;
-    error: string;
-  };
+  bundle?: { processing: boolean; code: string; error: string };
   updateCell: (id: string, content: string) => void;
 }
 
@@ -29,6 +25,7 @@ interface PreviewProps {
   code: string;
   status: string;
   isResizing: boolean;
+  cellId: string;
 }
 
 interface CodeEditorProps {
@@ -37,15 +34,34 @@ interface CodeEditorProps {
   runCode: () => void;
 }
 
-const Preview = ({ code, status, isResizing }: PreviewProps) => {
-  const iframe = useRef<any>();
+const Preview = ({ code, status, isResizing, cellId }: PreviewProps) => {
+  const iframe = useRef<HTMLIFrameElement>(null);
+  const [ready, setReady] = useState(false);
 
   useEffect(() => {
-    iframe.current.srcdoc = HTML;
-    setTimeout(() => {
-      iframe.current.contentWindow.postMessage(code, '*');
-    }, 50);
-  }, [code]);
+    const listener = (event: MessageEvent) => {
+      if (event.data?.type === 'ready' && event.data.cellId === cellId) {
+        setReady(true);
+      }
+    };
+    window.addEventListener('message', listener);
+    return () => window.removeEventListener('message', listener);
+  }, [cellId]);
+
+  // Reset iframe HTML whenever cell changes
+  useEffect(() => {
+    if (iframe.current) {
+      setReady(false);
+      iframe.current.srcdoc = getHTML(cellId);
+    }
+  }, [cellId]);
+
+  // Only send code once iframe is ready
+  useEffect(() => {
+    if (iframe.current && ready) {
+      iframe.current.contentWindow?.postMessage(code, '*');
+    }
+  }, [code, ready]);
 
   return (
     <div className={`${scssObj.baseClass}__preview-wrapper`}>
@@ -55,7 +71,6 @@ const Preview = ({ code, status, isResizing }: PreviewProps) => {
         title={`${scssObj.baseClass}__code-preview`}
         ref={iframe}
         sandbox="allow-scripts"
-        srcDoc={HTML}
       />
       {status && <div className={`${scssObj.baseClass}__preview-error`}>{status}</div>}
     </div>
@@ -66,18 +81,18 @@ const CodeEditor = ({ initialValue, onChange, runCode }: CodeEditorProps) => {
   const editorRef = useRef<any>();
   const dispatch = useDispatch();
 
-  const onEditorDidMount: EditorDidMount = (getValue: any, editor: any) => {
+  const handleEditorMount: OnMount = (editor, monaco) => {
     editorRef.current = editor;
 
     editor.onDidChangeModelContent(() => {
-      onChange(getValue());
+      onChange(editor.getValue());
     });
 
     editor.getModel()?.updateOptions({ tabSize: 2 });
   };
 
   const onFormatClick = () => {
-    const unFormatted = editorRef.current.getModel().getValue();
+    const unFormatted = editorRef.current.getValue();
     const formatted = prettier
       .format(unFormatted, {
         parser: 'babel',
@@ -91,7 +106,7 @@ const CodeEditor = ({ initialValue, onChange, runCode }: CodeEditorProps) => {
   };
 
   const onCopyClick = () => {
-    dispatch(copyText.request({ text: editorRef.current.getModel().getValue() }));
+    dispatch(copyText.request({ text: editorRef.current.getValue() }));
   };
 
   return (
@@ -121,11 +136,11 @@ const CodeEditor = ({ initialValue, onChange, runCode }: CodeEditorProps) => {
         />
       </div>
       <Editor
-        editorDidMount={onEditorDidMount}
+        onMount={handleEditorMount}
         value={initialValue}
         height="100%"
         language="javascript"
-        theme="dark"
+        theme="vs-dark"
         options={{
           wordWrap: 'on',
           minimap: { enabled: false },
@@ -145,14 +160,11 @@ const CodeCell = ({ cell, bundle, updateCell }: Props) => {
   const cumulativeCode = useCumulativeCode(cell.id);
   const [runCode, setRunCode] = useState(false);
   const [isResizing, setResizing] = useState(false);
-
   const dispatch = useDispatch();
 
   const bundler = async () => {
     dispatch(createBundle.request({ cellId: cell.id, input: cumulativeCode }));
-
     const result = await Bundle(cumulativeCode);
-
     dispatch(createBundle.success({ cellId: cell.id, bundle: result }));
   };
 
@@ -171,7 +183,6 @@ const CodeCell = ({ cell, bundle, updateCell }: Props) => {
     const timer = setTimeout(async () => {
       bundler();
     }, 750);
-
     // eslint-disable-next-line consistent-return
     return () => {
       clearTimeout(timer);
@@ -198,13 +209,31 @@ const CodeCell = ({ cell, bundle, updateCell }: Props) => {
                 onChange={(value) => updateCell(cell.id, value)}
               />
             </Resizable>
-            <div className={`${scssObj.baseClass}__progress-wrapper`}>
-              {!bundle || bundle.processing ? (
-                <div className={`${scssObj.baseClass}__progress-cover`}>
+            <div
+              className={`${scssObj.baseClass}__progress-wrapper`}
+              style={{ position: 'relative' }}
+            >
+              <Preview
+                isResizing={isResizing}
+                code={bundle?.code || ''}
+                status={bundle?.error || ''}
+                cellId={cell.id}
+              />
+              {(!bundle || bundle.processing) && (
+                <div
+                  className={`${scssObj.baseClass}__progress-cover`}
+                  style={{
+                    position: 'absolute',
+                    inset: 0,
+                    background: 'rgba(255,255,255,0.7)',
+                    display: 'flex',
+                    alignItems: 'center',
+                    justifyContent: 'center',
+                    zIndex: 10,
+                  }}
+                >
                   <LoadingSpinner type={SpinnerType.CubeFlipSpinner} />
                 </div>
-              ) : (
-                <Preview isResizing={isResizing} code={bundle.code} status={bundle.error} />
               )}
             </div>
           </>
