@@ -1,13 +1,14 @@
 /* eslint-disable jsx-a11y/alt-text */
-import { withRouter } from 'react-router-dom';
+import { useLocation } from 'react-router-dom';
 import scssObj from './_NotFound.scss';
 import Astronaut from '../animations/astronaut';
 import { ASSETS_BASE_URL } from '../../utils/assets';
 import Button from '../../elements/button';
 import Link from '../../elements/link';
 
-const NotFound = withRouter((history) => {
-  const state = (history.location.state || { from: undefined }) as { from?: string };
+const NotFound = () => {
+  const location = useLocation();
+  const state = (location.state || { from: undefined }) as { from?: string };
 
   return (
     <div className={scssObj.baseClass}>
@@ -30,6 +31,6 @@ const NotFound = withRouter((history) => {
       </div>
     </div>
   );
-});
+};
 
 export default NotFound;
