@@ -1,5 +1,7 @@
-import { RouteComponentProps } from 'react-router';
+// eslint-disable-next-line import/no-extraneous-dependencies
+import { createBrowserHistory } from 'history';
 
+export const history = createBrowserHistory();
 export interface SagaContext {
-  routerHistory: RouteComponentProps['history'];
+  routerHistory: typeof history;
 }
